P2+P3: 10/12-bit + full profile 0-3 support, Frame/PlaneData API
Adds VP9 profile 2 (10/12-bit 4:2:0) AND, as a side effect of P1's
already-general subsampling, profile 3 (10/12-bit non-4:2:0) -- the whole
official sweep is now 315/315 (all vp91/vp92/vp93 vectors), both SIMD-on
and forced-scalar, with the 304 8-bit vectors still byte-identical.

- framebuffer::Plane is u16-backed (8-bit samples widened, one code path
  for all depths); narrowing to u8 only at output via Plane::crop_u8.
- Public API: enum PlaneData { U8(Vec<u8>), U16(Vec<u16>) }; Frame gains
  bit_depth / subsampling_x / subsampling_y and y/u/v: PlaneData. 8-bit
  frames stay U8 (no memory bloat). DecodeError::UnsupportedBitDepth and
  every bit_depth==8 assert removed.
- loop_filter: bit_depth threaded through; the 8-bit constants
  (limit/blimit/thresh, flat-mask threshold, narrow_filter's 0x80 base and
  clamp4 range) now scale << (bit_depth-8) per spec 8.8.4/8.8.5 -- identity
  at 8-bit, so 8-bit output is unchanged. wide_filter (pure averaging)
  needed no change.
- LANDMINE closed: the loop-filter AVX2 kernel dispatch is now gated on
  bit_depth==8 (it was only guarded by the old 8-bit rejection); both AVX2
  kernels were also moved to u16 loads (cvtepu16/packus) while keeping
  8-bit-value arithmetic, so SIMD stays bit-exact under u16 storage.
- Tests: i420_bytes emits U16 as 2 LE bytes/sample (libvpx high-depth md5
  convention), making vp92/vp93 MD5-checkable; all frame.y/u/v consumers
  (sweep/bench/synthetic/api/example) adapted to PlaneData.
- 152 tests, conformance 12/12 both SIMD modes, ffmpeg cross-decode 8/8,
  clippy 3 baseline, fmt clean.

+++ src/predict.rs
@@ -273,7 +273,7 @@ pub fn predict_intra(
 
     for i in 0..size {
         for j in 0..size {
-            plane.set(x + j, y + i, pred[at(i, j)] as u8);
+            plane.set(x + j, y + i, pred[at(i, j)] as u16);
         }
     }
 }
@@ -484,9 +484,9 @@ fn block_inter_predict(
     // for the horizontal pass's column, `r` alone for the vertical pass's row -- see
     // `simd.rs`'s doc comment for the derivation). Falls through to the scalar loop for:
     // the scaled path (rarer, harder to vectorize -- kept scalar per this wave's scope),
-    // bit_depth != 8 (never actually reached today, lib.rs rejects non-8-bit streams
-    // before decode, but the scalar loop handles it generically so the guard is free to
-    // keep), width 4 (not a multiple of the AVX2 kernel's 8-wide lane group), and any
+    // bit_depth != 8 (10/12-bit frames: the AVX2 kernel's arithmetic -- clip1 to 0..=255 --
+    // is only valid at 8-bit; the scalar loop below handles every depth generically), width
+    // 4 (not a multiple of the AVX2 kernel's 8-wide lane group), and any
     // block whose source window would need the scalar path's per-pixel edge clamp
     // (border replication) -- only near reference-frame edges; replicating that with
     // AVX2 would need a byte gather, which x86 doesn't have.
@@ -618,7 +618,7 @@ pub fn predict_inter(
             } else {
                 preds[0][i * w + j]
             };
-            dst.set(x + j, y + i, value as u8);
+            dst.set(x + j, y + i, value as u16);
         }
     }
 }
@@ -632,7 +632,7 @@ mod tests {
         let mut p = Plane::new(width, height);
         for y in 0..height {
             for x in 0..width {
-                p.set(x, y, fill);
+                p.set(x, y, fill as u16);
             }
         }
         p
@@ -650,12 +650,12 @@ mod tests {
     fn v_pred_copies_above_row() {
         let mut plane = make_plane(8, 8, 0);
         for x in 0..4 {
-            plane.set(x, 3, 50 + x as u8);
+            plane.set(x, 3, 50 + x as u16);
         }
         predict_intra(&mut plane, 0, 4, false, true, false, TX4, V_PRED, 7, 7, 8);
         for x in 0..4 {
-            assert_eq!(plane.get(x, 4), 50 + x as u8);
-            assert_eq!(plane.get(x, 5), 50 + x as u8);
+            assert_eq!(plane.get(x, 4), 50 + x as u16);
+            assert_eq!(plane.get(x, 5), 50 + x as u16);
         }
     }
 
@@ -663,12 +663,12 @@ mod tests {
     fn h_pred_copies_left_col() {
         let mut plane = make_plane(8, 8, 0);
         for y in 0..4 {
-            plane.set(3, y, 60 + y as u8);
+            plane.set(3, y, 60 + y as u16);
         }
         predict_intra(&mut plane, 4, 0, true, false, false, TX4, H_PRED, 7, 7, 8);
         for y in 0..4 {
-            assert_eq!(plane.get(4, y), 60 + y as u8);
-            assert_eq!(plane.get(5, y), 60 + y as u8);
+            assert_eq!(plane.get(4, y), 60 + y as u16);
+            assert_eq!(plane.get(5, y), 60 + y as u16);
         }
     }
